* WebSocket10.cs: Cosmetic
* HttpUtility.cs: Correct a bug when generating a QueryString

* HttpRouting.cs: Use int type when :int is use in the route

* HttpClient.cs:
* HttpClientRequest.cs: Add Async support for HttpClient

+++ Erasme.Http/Erasme.Http/HttpRouting.cs
@@ -84,8 +84,8 @@ namespace Erasme.Http
 
 								if (paramType == "int")
 								{
-									long paramValue;
-									if (Int64.TryParse(p, out paramValue))
+									int paramValue;
+									if (Int32.TryParse(p, out paramValue))
 										value = paramValue;
 									else
 										break;
